Se corrigio un el tipo de dato de un atributo en la entidad cuentas amazon

+++ src/main/java/lyjstoreapirest/demo/cuentaAmazon/domain/model/CuentaAmazon.java
@@ -22,7 +22,7 @@ public class CuentaAmazon {
     private String direccion2;
     @Column(name = "codigo_postal")
     private Integer codigoPostal;
-    private Integer telefono;
+    private Double telefono;
     private String ciudad;
     private String estado;
 }
